refactor: move logger to separate file as requested in PR feedback
- Create new packages/xmtp/src/lib/logger.ts file
- Move logger implementation from jwt.ts to dedicated logger.ts
- Update exports in index.ts to export logger from ./lib/logger
- Update all imports across codebase to use new logger location
- Maintain conditional logging behavior based on XMTP_DEBUG
- Address GitHub PR comment: 'This isn't a jwt thing'
